Document and improve json extension
